+++ sickrage/core/__init__.py
@@ -95,8 +95,6 @@ class Core(object):
         self.newest_version = None
         self.newest_version_string = None
 
-        self.api_url = None
-
         self.user_agent = 'SiCKRAGE.CE.1/({};{};{})'.format(platform.system(), platform.release(), str(uuid.uuid1()))
         self.sys_encoding = get_sys_encoding()
         self.languages = [language for language in os.listdir(sickrage.LOCALE_DIR) if '_' in language]
@@ -192,9 +190,6 @@ class Core(object):
 
         # set language
         self.config.change_gui_lang(self.config.gui_lang)
-
-        # set api url
-        self.api_url = 'https://api.sickrage.ca/v1/'
 
         # set socket timeout
         socket.setdefaulttimeout(self.config.socket_timeout)